+++ libsolidity/formal/SMTChecker.cpp
@@ -71,23 +71,38 @@ bool SMTChecker::visit(FunctionDefinition const& _function)
 			_function.location(),
 			"Assertion checker does not yet support constructors and functions with modifiers."
 		);
-	m_currentFunction = &_function;
-	m_interface->reset();
-	m_pathConditions.clear();
+	if (isRootFunction())
+	{
+		m_interface->reset();
+		m_pathConditions.clear();
+		m_expressions.clear();
+		resetStateVariables();
+		initializeLocalVariables(_function);
+	}
+	else
+	{
+		initializeFunctionParameters(_function);
+	}
+
+	declareFunction(_function);
+	m_currentFunction.push_back(&_function);
 	m_loopExecutionHappened = false;
-	resetStateVariables();
-	initializeLocalVariables(_function);
 	return true;
 }
 
-void SMTChecker::endVisit(FunctionDefinition const&)
+void SMTChecker::endVisit(FunctionDefinition const& _function)
 {
-	// TODO we could check for "reachability", i.e. satisfiability here.
-	// We only handle local variables, so we clear at the beginning of the function.
-	// If we add storage variables, those should be cleared differently.
-	removeLocalVariables();
-	m_currentFunction = nullptr;
+	if (_function.returnParameters().size())
+		if (hasVariable(*(_function.returnParameters()[0])))
+			m_functionReturn.push_back(currentValue(*(_function.returnParameters()[0])));
+	//removeLocalVariables(*m_currentFunction.back());
+	m_currentFunction.pop_back();
+	if (isRootFunction())
+		removeLocalVariables();
+	else
+		m_functionArguments.pop_back();
 }
+
 
 bool SMTChecker::visit(IfStatement const& _node)
 {
@@ -240,7 +255,7 @@ void SMTChecker::endVisit(TupleExpression const& _tuple)
 	if (_tuple.isInlineArray() || _tuple.components().size() != 1)
 		m_errorReporter.warning(
 			_tuple.location(),
-			"Assertion checker does not yet implement tules and inline arrays."
+			"Assertion checker does not yet implement tuples and inline arrays."
 		);
 	else
 		defineExpr(_tuple, expr(*_tuple.components()[0]));
@@ -365,6 +380,66 @@ void SMTChecker::endVisit(FunctionCall const& _funCall)
 		checkBooleanNotConstant(*args[0], "Condition is always $VALUE.");
 		addPathImpliedExpression(expr(*args[0]));
 	}
+	else if (funType.kind() == FunctionType::Kind::Internal ||
+			 funType.kind() == FunctionType::Kind::External
+			)
+	{
+		FunctionDefinition const* _funDef;
+		if (Identifier const* _fun = dynamic_cast<Identifier const*>(&_funCall.expression()))
+			_funDef = dynamic_cast<FunctionDefinition const*>(_fun->annotation().referencedDeclaration);
+		else if (MemberAccess const* _fun = dynamic_cast<MemberAccess const*>(&_funCall.expression()))
+			_funDef = dynamic_cast<FunctionDefinition const*>(_fun->annotation().referencedDeclaration);
+		else
+		{
+			m_errorReporter.warning(
+				_funCall.location(),
+				"Assertion checker does not yet implement this type of function call."
+			);
+			return;
+		}
+
+		if (visitedFunction(_funDef))
+			m_errorReporter.warning(
+					_funCall.location(),
+					"Assertion checker does not support recursive function calls.",
+					SecondarySourceLocation().append("Starting from function:", _funDef->location())
+				);
+		else if (_funDef && _funDef->isImplemented())
+		{
+			auto callArgs = _funCall.arguments();
+			auto params = _funDef->parameters();
+			solAssert(callArgs.size() == params.size(), "");
+			vector<smt::Expression> funArgs;
+			for (auto arg: callArgs)
+				funArgs.push_back(expr(*arg));
+			m_functionArguments.push_back(funArgs);
+			_funDef->accept(*this);
+			if (_funDef->returnParameters().size())
+			{
+				vector<smt::Expression> paramExprs;
+				for (auto _p: params)
+					paramExprs.push_back(currentValue(*_p));
+				//defineExpr(_funCall, m_functionReturn.back());
+				defineExpr(_funCall, m_functions.at(_funDef)(funArgs));
+				m_interface->addAssertion(currentValue(*_funDef->returnParameters()[0]) == m_functions.at(_funDef)(paramExprs));
+				m_functionReturn.pop_back();
+			}
+		}
+		else
+		{
+			m_errorReporter.warning(
+				_funCall.location(),
+				"Assertion checker does not support calls to functions without implementation."
+			);
+		}
+	}
+	else
+	{
+		m_errorReporter.warning(
+			_funCall.location(),
+				"Assertion checker does not yet implement this type of function call."
+		);
+	}
 }
 
 void SMTChecker::endVisit(Identifier const& _identifier)
@@ -410,6 +485,12 @@ void SMTChecker::endVisit(Literal const& _literal)
 			_literal.annotation().type->toString() +
 			")."
 		);
+}
+
+void SMTChecker::endVisit(Return const& _return)
+{
+	if (hasExpr(*_return.expression()))
+		m_interface->addAssertion(expr(*_return.expression()) == newValue(*m_currentFunction.back()->returnParameters()[0]));
 }
 
 void SMTChecker::arithmeticOperation(BinaryOperation const& _op)
@@ -578,7 +659,7 @@ void SMTChecker::checkCondition(
 
 	vector<smt::Expression> expressionsToEvaluate;
 	vector<string> expressionNames;
-	if (m_currentFunction)
+	if (m_currentFunction.size())
 	{
 		if (_additionalValue)
 		{
@@ -607,7 +688,7 @@ void SMTChecker::checkCondition(
 	{
 		std::ostringstream message;
 		message << _description << " happens here";
-		if (m_currentFunction)
+		if (m_currentFunction.size())
 		{
 			std::ostringstream modelMessage;
 			modelMessage << "  for:\n";
@@ -723,6 +804,25 @@ smt::CheckResult SMTChecker::checkSatisfiable()
 	return checkSatisfiableAndGenerateModel({}).first;
 }
 
+void SMTChecker::initializeFunctionParameters(FunctionDefinition const& _function)
+{
+	auto const& funParams = _function.parameters();
+	auto const& callArgs = m_functionArguments.back();
+	solAssert(funParams.size() == callArgs.size(), "");
+	for (unsigned i = 0; i < funParams.size(); ++i)
+		if (createVariable(*funParams[i]))
+			m_interface->addAssertion(callArgs[i] == newValue(*funParams[i]));
+
+	for (auto const& variable: _function.localVariables())
+		if (createVariable(*variable))
+			setZeroValue(*variable);
+
+	if (_function.returnParameterList())
+		for (auto const& retParam: _function.returnParameters())
+			if (createVariable(*retParam))
+				setZeroValue(*retParam);
+}
+
 void SMTChecker::initializeLocalVariables(FunctionDefinition const& _function)
 {
 	for (auto const& variable: _function.localVariables())
@@ -738,6 +838,33 @@ void SMTChecker::initializeLocalVariables(FunctionDefinition const& _function)
 			if (createVariable(*retParam))
 				setZeroValue(*retParam);
 }
+
+void SMTChecker::removeLocalVariables()
+{
+	for (auto it = m_variables.begin(); it != m_variables.end(); )
+	{
+		if (it->first->isLocalVariable())
+			it = m_variables.erase(it);
+		else
+			++it;
+	}
+}
+
+/*
+void SMTChecker::removeLocalVariables(FunctionDefinition const& _function)
+{
+	for (auto const& variable: _function.localVariables())
+		m_variables.erase(m_variables.find(variable));
+
+	for (auto const& param: _function.parameters())
+		m_variables.erase(m_variables.find(&*param));
+
+	if (_function.returnParameterList())
+		for (auto const& retParam: _function.returnParameters())
+			if (hasVariable(*retParam))
+				m_variables.erase(m_variables.find(&*retParam));
+}
+*/
 
 void SMTChecker::resetStateVariables()
 {
@@ -781,7 +908,7 @@ bool SMTChecker::createVariable(VariableDeclaration const& _varDecl)
 {
 	if (SSAVariable::isSupportedType(_varDecl.type()->category()))
 	{
-		solAssert(m_variables.count(&_varDecl) == 0, "");
+		solAssert(m_variables.count(&_varDecl) == 0 || !isRootFunction(), "");
 		m_variables.emplace(&_varDecl, SSAVariable(_varDecl, *m_interface));
 		return true;
 	}
@@ -838,40 +965,46 @@ void SMTChecker::setUnknownValue(VariableDeclaration const& _decl)
 
 smt::Expression SMTChecker::expr(Expression const& _e)
 {
-	if (!m_expressions.count(&_e))
+	if (!hasExpr(_e))
 	{
 		m_errorReporter.warning(_e.location(), "Internal error: Expression undefined for SMT solver." );
 		createExpr(_e);
 	}
-	return m_expressions.at(&_e);
+	return prev(m_expressions.upper_bound(&_e))->second;
+}
+
+bool SMTChecker::hasExpr(Expression const& _e) const
+{
+	return m_expressions.count(&_e);
+}
+
+bool SMTChecker::hasVariable(VariableDeclaration const& _var) const
+{
+	return m_variables.count(&_var);
 }
 
 void SMTChecker::createExpr(Expression const& _e)
 {
-	if (m_expressions.count(&_e))
-		m_errorReporter.warning(_e.location(), "Internal error: Expression created twice in SMT solver." );
-	else
+	solAssert(_e.annotation().type, "");
+	string exprSymbol = uniqueSymbol(_e) + "_" + to_string(m_expressions.count(&_e));
+	switch (_e.annotation().type->category())
 	{
-		solAssert(_e.annotation().type, "");
-		switch (_e.annotation().type->category())
-		{
-		case Type::Category::RationalNumber:
-		{
-			if (RationalNumberType const* rational = dynamic_cast<RationalNumberType const*>(_e.annotation().type.get()))
-				solAssert(!rational->isFractional(), "");
-			m_expressions.emplace(&_e, m_interface->newInteger(uniqueSymbol(_e)));
-			break;
-		}
-		case Type::Category::Address:
-		case Type::Category::Integer:
-			m_expressions.emplace(&_e, m_interface->newInteger(uniqueSymbol(_e)));
-			break;
-		case Type::Category::Bool:
-			m_expressions.emplace(&_e, m_interface->newBool(uniqueSymbol(_e)));
-			break;
-		default:
-			solUnimplementedAssert(false, "Type not implemented.");
-		}
+	case Type::Category::RationalNumber:
+	{
+		if (RationalNumberType const* rational = dynamic_cast<RationalNumberType const*>(_e.annotation().type.get()))
+			solAssert(!rational->isFractional(), "");
+		m_expressions.emplace(&_e, m_interface->newInteger(exprSymbol));
+		break;
+	}
+	case Type::Category::Address:
+	case Type::Category::Integer:
+		m_expressions.emplace(&_e, m_interface->newInteger(exprSymbol));
+		break;
+	case Type::Category::Bool:
+		m_expressions.emplace(&_e, m_interface->newBool(exprSymbol));
+		break;
+	default:
+		solUnimplementedAssert(false, "Type not implemented.");
 	}
 }
 
@@ -909,13 +1042,58 @@ void SMTChecker::addPathImpliedExpression(smt::Expression const& _e)
 	m_interface->addAssertion(smt::Expression::implies(currentPathConditions(), _e));
 }
 
-void SMTChecker::removeLocalVariables()
+bool SMTChecker::isRootFunction()
 {
-	for (auto it = m_variables.begin(); it != m_variables.end(); )
+	return m_currentFunction.size() == 0;
+}
+
+bool SMTChecker::visitedFunction(FunctionDefinition const* _funDef)
+{
+	return find(m_currentFunction.begin(), m_currentFunction.end(), _funDef) != m_currentFunction.end();
+}
+
+void SMTChecker::declareFunction(FunctionDefinition const& _fun)
+{
+	// If the function does not return anything we just inline
+	// the contents.
+	// If the function returns something we create an expression.
+	unsigned retSize = _fun.returnParameters().size();
+	if (retSize >  1)
 	{
-		if (it->first->isLocalVariable())
-			it = m_variables.erase(it);
-		else
-			++it;
+		m_errorReporter.warning(
+			_fun.location(),
+			"Assertion checker does not yet support functions with multiple return values."
+		);
+	}
+	else if (retSize == 1)
+	{
+		vector<smt::Sort> paramSorts;
+		for (auto _param: _fun.parameters())
+			paramSorts.push_back(SMTSort(*_param->type()));
+
+		m_functions.emplace(&_fun, m_interface->newFunction(_fun.name(), paramSorts, SMTSort(*_fun.returnParameters()[0]->type())));
+	}
+}
+
+smt::Sort SMTChecker::SMTSort(Type const& _type)
+{
+	return SMTSort(_type.category());
+}
+
+smt::Sort SMTChecker::SMTSort(Type::Category const& _category)
+{
+	switch (_category)
+	{
+		case Type::Category::Address:
+		case Type::Category::Integer:
+		case Type::Category::RationalNumber:
+		case Type::Category::Enum:
+		case Type::Category::FixedBytes:
+			return smt::Sort::Int;
+		case Type::Category::Bool:
+			return smt::Sort::Bool;
+		default:
+			// TODO error instead of assert
+			solAssert(false, "");
 	}
 }